+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 0.0.7 - 2020-10-11
+
+ * Fixed Multiline comment detection
+ * Added before and after statement support
+
+
+
+## 0.0.6 - 2020-10-11
+
+ * Added outline for livecodescript
+ * Severanl performance improvements
+
+
 ## 0.0.5 - 2020-10-07
 
  * UPDATED Command and function names are case insensitive in lcs now
@@ -12,7 +25,7 @@
  * ADDED Basic support for extension "lc" (thanks to SparkOut from lc forum)
  * ADDED Folding to "if" statement (LCB)
 
-## 0.0.3 - 2020-10-05
+## 0.0.3 - 2020-10-11
 
  * ADDED   LCB handler folding
  * FIXED   Recognizing . as part of command names